refactor: ok, error에 관한 responseDto 하나로 통일
