--- src/main/java/chess/domain/dto/ErrorResponseDto.java
@@ -1,24 +1,0 @@
-package chess.domain.dto;
-
-public class ErrorResponseDto {
-
-    private final ResponseStatus status;
-    private final String message;
-
-    private ErrorResponseDto(ResponseStatus status, String message) {
-        this.status = status;
-        this.message = message;
-    }
-
-    public static ErrorResponseDto message(String message) {
-        return new ErrorResponseDto(ResponseStatus.ERROR, message);
-    }
-
-    public String getStatus() {
-        return status.getStatus();
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
